Create a packet-based connection between the testing server and client
Instead of using TCP, we use a custom, testing Read/Write implementation based on two channels.
It's not clear if this is appropriate to implement DTLS transport layer, but there doesn't seem to be any other option.
This approach also currently fails when there is an error during the handshake, the test then hangs forever. It could be caused by not propagating Channel drop properly to the other side.